backend/app/core/middleware.py
import time
from starlette.middleware.base import BaseHTTPMiddleware
from app.core.logging import logger

class LogRequestsMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request, call_next):
        start_time = time.time()
        response = await call_next(request)
        duration = time.time() - start_time
        
        logger.info(
            "http_request",
            method=request.method,
            path=request.url.path,
            status_code=response.status_code,
            duration=f"{duration:.4f}s"
        )
        return response
