backend/pyproject.toml
[project]
name = "institute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "boto3>=1.42.91",
    "celery>=5.6.3",
    "cryptography>=46.0.7",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "redis>=7.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.49",
    "structlog>=25.5.0",
    "uvicorn>=0.44.0",
]
